remove comments presenter
